Better separate header contents by functionality.
