+++ libtk205/include/typeinfo_205.h
@@ -1,0 +1,16 @@
+#ifndef TYPEINFO_205_H_
+#define TYPEINFO_205_H_
+
+#include <string_view>
+
+namespace ASHRAE205_NS {
+
+    struct enum_info
+    { 
+        std::string_view enumerant_name; 
+        std::string_view display_text; 
+        std::string_view description;
+    };
+}
+
+#endif // TYPEINFO_205_H_
