+++ libtk205/src/error_handling_tk205.cpp
@@ -13,7 +13,7 @@ namespace ASHRAE205_NS {
 
     void Show_message(msg_severity severity, const std::string &message)
     {
-        static std::map<msg_severity, std::string> severity_str {
+        static std::map<msg_severity, std::string_view> severity_str {
             {msg_severity::DEBUG_205, "DEBUG"},
             {msg_severity::INFO_205, "INFO"},
             {msg_severity::WARN_205, "WARN"},
